rename function, move marshalling to own file
